Implement script to fully convert and clean pdf
