+++ utils/pdf_to_text.sh
@@ -1,0 +1,3 @@
+java -jar $PDFBOX_HOME ExtractText $1 "$1.tmp"
+./clean_txt.py -o $2 $1.tmp
+rm $1.tmp
